requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
transformers
